Refactored Dialogue speak function.
Now the talking character is send as parameter of the signal, and the
color of the dialog text is defined in that character.
Characters can have a spoke function that is called when their dialog
finishes.
When a character talks, a bubble appears beneath her.
Fixed some bugs in the Intro behavior.
Added the Data script to handle a data registry across all the scenes
and objects in the game.