Fixed camera zoom bug on game start.
Fixed camera zoom change when leaving/entering a zone while the previous
animation were still running.
Updated TalkingBubble so it uses a RichTextLabel to show a wave
animation for its dots.
Updated TalkingBubble to remove and append the
wave-dot-animated-characters on dissapear and appear.
Added TileSet atlasses for placing the sub-god statues in the Temple.