Dialog diagram can now have a return node to show the last dialog menu.
Dialog nodes can now turn off and on dialog menu options by their ID (slot index for the time being).
When the DialogMenu is requested, the player can't move Teotriste.
Added an Event signal to toggle Teotriste movement.
Ready to work on fixing the problem in the TalkingBubble behavior and continue to add functionality to the Dialog system.