Improved how dialogs can be skipped with Z.
Now the Dialog scene listens for the event hud_accept_pressed and delegates part of the flow to the Autofill node.
Once the Autofill has completed, the Dialog scene decides what should be done based on the dialog type.
Skipping the intro was also refactored, there are still weird behaviors. Nothing that can't be fixed later.

[ Ready to work on Talk action and allowing to turn off dialog options foreva ]