Added surface types to Data singleton.
Surface class now has an export to define the type of the surface.
Hook is sent back when the state HOLD is not in a surface of type WATER.
Hook is sent back after 1 sec of not detecting a surface in HOLD state.
Minor adjustments on how the hook is placed in FISHPREPARE and THROW.
When Hook goes to IDLE state, it resets its position.