Fishing surfaces can now have specific and sacred fishes.
Added a function to FishingDatabase which get a fish by its type.
The type of fishes for the fishing surface are set in an array. It will take a random fish from that array.
Only one sacred fish can inhabit a fishing surface. For the momento.
Added two utility methods to Utils.gd:
- One for getting a random index position from an array.
- One for getting a random element from an array.
< Closes $17k and $17o > /:</