Finished first iteration of fish tank
Added the foreground layer for the fish tank.
Added the columns, the arch and their collisions.
Created an Area2D to handle the change of the z-index for columns and
arch.
Created the collision that wont allow players to enter the tank.
Now the footstep SFX is handled by the FootArea in the player.
Updated the map bushes sprite.
< Closes card $16v >