Improvements in FishingSurface functionality.
Now the surface can be inside a Node.
FishingSurface will not use different CollisionPolygon2D to spawn
fishes. In order to spawn fishes in different areas, a FishingSurface
must be created for each spawning area.
Fish tank fishing surfaces are now in the FishTank scene.
Fixed a bug related with the position of the hook inside the fishing
surface. This improves how the fish examining shadow and the hook
position are calculated during fishing.
Only fishing surfaces with sacred fishes render their list of fishes in
the DebugOverlay.
< Closes card $18a and $18c >