Finished collisions for front columns in altar
Separated the right-front-column drebis to better handle its collisions
and to have a mask to hide Teotriste to simulate it is behind the small
rocks on the ground.
Added some collisions to the main structure in the altar so the player
can traverse it.
--- Working on enabling/disabling the main structure collisions based on
the position of Teotriste in relation to it ---
< Progress on card $1ii >