Fixed bug on continuing dialog after selecting > 0 option
Selecting the B option on the first menu was making the game crash
because the _selected_slot variable wasn't restarting to 0, so the next
node in the graph was trying to load the NID for that slot.
Fixed a bug in the NID 5 of RocbertoIntro because "Teotriste" was used as
the name of the actor. Changed to "Player".