Updated Sacred Fire Altar activation animation
Now the animation doesn't activate the temple entrance button, instead,
it calls the method in TempleEntrance.gd that plays the temple
activation animation and listens for its _activated_ signal to continue
with the camera movements.
The camera movement times are now exported in AltarActivation.gd.
Exported the assets that will be needed to create the activated version
of the temple entrance.
Added the door to the temple entrance with its collisions.

< Progress on $1uk >