Updated ZIndexChanger for Platform back columns
Made some improvements on the ZIndexChanger node. Now it can disable
other ZIndexChangers when Enter, Exit or Both.
Updated the shapes of the polygons in both front and back columns in the
platform to get a better behavior of collisions and ZIndexChangers.
Added a modulation to the FoorArea in the player to facilitate debugging
collisions.

< Progress in $1ue >