Created floor holes colitions.
Created a function to force the VisibilityDisabler to modify all its
parent's children's Area2D, PhysicsBody2D and Light2D despite some of
them are in the array of nodes to ignore. I need to make a better
approach to this.
Put the sacred fish near the Sacred Fire altar again for testing the
temple activation animation.

< Progress in $1uv >