Updated PopupMenu in custom OptionButton
This allowed us to replace the font and radio button images for the list
of positions to teletransport to in each level.
Now the list of positions appears in the top of the pause menu and has a
label.
Added all the available zone positions for Chapter0.1 with their English
names because yes.