Updated the assets for aged fingers and platform.
Updated the floor asset to remove the platform cover.
The platform cover in the floor now animates independently.
Updated the mask for the fingers. NEEDS A FIX.
Camera zooms out just before starting the temple activation. Once the
activation has finished, it zooms to its initial value after moving back
to the Player.

< Progress in $1z3 >