Updated all sprites for aged temple entrance
Created node for making a Node2D have a floating VFX.
Updated masks for front pillars and fingers.
Fixed a bug in the door button press animation.
Created a Area2D for the temple in order to zoom out and offset in Y the
player camera when she enters the entrance... and vice versa.
Updated facade torch animation values.
Game controls are ALWAYS shown in the build version.
Added functionality to the Chapter0.1 script (World.gd) to place the
player in a specific position in the scene on start. This position is
set via a NodePath.

< Closes $1z3 >