Aged part of the Camino del regender.
Bottom floor, pyramids and columns.
Updated the tip of the columns because the previous ones estaban como malucas.
Added foreground layer for the Camino del regender.
Updated the type of the speed_modifier in Surface.gd to be a Vector2
instead of a float. This will allow us to change the movement speed of
the PC in both axis separately.
< Progress in $1zh >